Refactor workflow event history pages to pages directory
